refactor: Remove unused fetchResultCategoryDetailInfo function and update import paths
