+++ app/ui/result/result-category-detail/result-category-detail.tsx
@@ -1,12 +1,9 @@
 'use client';
 import ResultCategoryDetailContent from '@/app/ui/result/result-category-detail-content/result-category-detail-content';
-import { fetchResultCategoryDetailInfo } from '@/app/business/services/result/result.query';
+import { useFetchResultCategoryDetailInfo } from '@/app/store/querys/result';
 import ResultCategoryDetailContentSkeleton from '@/app/ui/result/result-category-detail-content/result-category-detail-content.skeleton';
 import ResultCategoryDetailDialog from './result-category-detail-dialog';
 import { Suspense } from 'react';
-import { QUERY_KEY } from '@/app/utils/query/react-query-key';
-import { ResultCategoryDetailResponse } from '@/app/business/services/result/result.type';
-import { useSuspenseQuery } from '@tanstack/react-query';
 
 export default function ResultCategoryDetail({ category }: { category: string }) {
   return (
@@ -19,11 +16,7 @@ export default function ResultCategoryDetail({ category }: { category: string })
 }
 
 function ResultCategoryDetailInfo({ category }: { category: string }) {
-  const { data } = useSuspenseQuery<ResultCategoryDetailResponse>({
-    queryKey: [QUERY_KEY.CATEGORY],
-    staleTime: Infinity,
-    queryFn: () => fetchResultCategoryDetailInfo(category),
-  });
+  const { data } = useFetchResultCategoryDetailInfo(category);
 
   return <ResultCategoryDetailContent info={data} />;
 }
